requirements.txt
aniso8601==9.0.1
attrs==23.1.0
blinker==1.7.0
click==8.1.7
Flask==3.0.0
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.3
PyMySQL==1.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
referencing==0.32.0
rpds-py==0.16.2
SQLAlchemy==2.0.24
typing_extensions==4.9.0
Werkzeug==3.0.1
